requirements.txt
pydantic-settings>=2.8.1
uvicorn>=0.34.0
fastapi>=0.115.10
pydantic>=2.10.6
langchain>=0.3.19
python-dotenv>=1.0.1
langchain-community>=0.3.18
langchain-openai>=0.3.7
langchain-text-splitters>=0.3.6